+++ examples/pytorch/diffpool/model/dgl_layers/gnn.py
@@ -8,6 +8,7 @@ import dgl.function as fn
 from .aggregator import MaxPoolAggregator, MeanAggregator, LSTMAggregator
 from .bundler import Bundler
 from ..model_utils import masked_softmax
+from model.loss import EntropyLoss
 
 
 class GraphSageLayer(nn.Module):
@@ -118,22 +119,3 @@ class DiffPoolBatchedGraphLayer(nn.Module):
             self.loss_log[loss_name] = loss_layer(adj, adj_new, assign_tensor)
 
         return adj_new, h
-    
-
-
-
-
-class EntropyLoss(nn.Module):
-    # Return Scalar
-    def forward(self, adj, anext, s_l):
-        entropy = (torch.distributions.Categorical(probs=s_l).entropy()).sum(-1).mean(-1)
-        assert not torch.isnan(entropy)
-        return entropy
-
-
-class LinkPredLoss(nn.Module):
-
-    def forward(self, adj, anext, s_l):
-        link_pred_loss = (adj - s_l.matmul(s_l.transpose(-1, -2))).norm(dim=(1, 2))
-        link_pred_loss = link_pred_loss / (adj.size(1) * adj.size(2))
-        return link_pred_loss.mean()
